requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
pdfplumber>=0.10.0
python-dotenv>=1.0.0
gradio>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tiktoken>=0.5.0
sentence-transformers>=2.2.2
torch>=2.0.0
scikit-learn>=1.3.0
